[rename/WEDPRO-43] rename: 폴더 구조 변경
