+++ src/containers/gallery/GallerySupportContainer.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function GallerySupportContainer() {
+  return <div>123</div>;
+}
