+++ src/containers/gallery/GalleryWeddingContainer.jsx
@@ -1,7 +1,7 @@
-import React from 'react';
+import React from "react";
 //import {RiArrowDropLeftLine,RiArrowDropRightLine} from 'react-icons/ri';
 import { Swiper, SwiperSlide } from "swiper/react"; // basic
-import {EffectCoverflow, Navigation } from 'swiper';
+import { EffectCoverflow, Navigation } from "swiper";
 
 // Import Swiper styles
 
@@ -9,21 +9,16 @@ import "swiper/css";
 import "swiper/css/pagination";
 import "swiper/css/navigation";
 
-import "./styles.css";
-import GalleryWeddingBox from '../components/GalleryWeddingBox/GalleryWeddingBox';
+import "../style/styles.css";
+import GalleryWeddingBox from "../../components/GalleryWeddingBox/GalleryWeddingBox";
 
+const banners = [1, 2, 3, 4, 5, 6];
 
-const banners = [
-  1,2,3,4,5,6
-];
-
-
-
-export default function GalleryWedding() {  
+export default function GalleryWeddingContainer() {
   return (
     <>
       <Swiper
-        effect={'coverflow'}
+        effect={"coverflow"}
         slidesPerView={3} // 동시에 보여줄 슬라이드 갯수
         //spaceBetween={50}// 슬라이드간 간격
         //slidesPerGroup={2} //그룹으로 묶을 수
@@ -32,31 +27,25 @@ export default function GalleryWedding() {
         centeredSlides={true}
         pagination={{
           clickable: true,
-          el:'.swiper-pagination'          
-        }}        
+          el: ".swiper-pagination",
+        }}
         coverflowEffect={{
           rotate: 0,
           stretch: 100,
           depth: 200,
           modifier: 1.5,
-          slideShadows : false,
+          slideShadows: false,
         }}
-        modules={[ Navigation,EffectCoverflow]}
-        className='swiper-container two'
+        modules={[Navigation, EffectCoverflow]}
+        className="swiper-container two"
         grabCursor={true}
-        
-    
       >
-    
-      {
-        banners.map((value,idx)=>(
+        {banners.map((value, idx) => (
           <SwiperSlide key={idx}>
-            <GalleryWeddingBox className="swiper-image"/>
+            <GalleryWeddingBox className="swiper-image" />
           </SwiperSlide>
-        ))
-      }
-      </Swiper>    
-  
-  </>
-  )
+        ))}
+      </Swiper>
+    </>
+  );
 }
