+++ src/containers/goods/GoodsSupportContainer.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function GoodsSupportContainer() {
+  return <GoodsSupportContainer />;
+}
