--- src/pages/GallerySupport.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function GallerySupport() {
-  return (
-    <div>GallerySupport</div>
-  )
-}
