--- src/pages/GoodsSupport.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function GoodsSupport() {
-  return (
-    <div>GoodsSupport</div>
-  )
-}
