+++ src/pages/gallery/GallerySupport.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import GallerySupportContainer from "../../containers/gallery/GallerySupportContainer";
+
+export default function GallerySupport() {
+  return <GallerySupportContainer />;
+}
