+++ src/pages/gallery/GalleryWedding.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import GalleryWeddingContainer from "../../containers/gallery/GalleryWeddingContainer";
+
+export default function GalleryWedding() {
+  return <GalleryWeddingContainer />;
+}
