+++ src/pages/goods/GoodsProduct.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import GoodsProductContainer from "../../containers/goods/GoodsProductContainer";
+
+export default function GoodsProduct() {
+  return <GoodsProductContainer />;
+}
