+++ src/pages/goods/GoodsSupport.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import GoodsSupportContainer from "../../containers/goods/GoodsSupportContainer";
+
+export default function GoodsSupport() {
+  return <GoodsSupportContainer />;
+}
